Nivel novato Super Trunfo Nível novato concluído, passando para nível aventureiro e preparando para fazer grandes mudanças no código.

+++ CartasSuperTrunfo.c
@@ -4,14 +4,16 @@ int main(){
 
 printf ("Bem vindo ao desafio Super Trunfo! \n");
 
+//variáveis das cartas a serem cadastradas
 char nome[20];
 char carta[20];
 char cidade[20];
-float populacao;
+int populacao;
 float area;
 float pib;
 int pontostur;
 
+//cadastro da CARTA 01 -------------------------------------------
 printf("Digite o nome do estado da primeira carta: \n");
 scanf("%s", &nome);
 
@@ -22,7 +24,7 @@ printf("Nome da cidade: \n");
 scanf("%s", &carta);
 
 printf("População: \n");
-scanf("%f", &populacao);
+scanf("%d", &populacao);
 
 printf("Área em KM2: \n");
 scanf("%f", &area);
@@ -31,7 +33,7 @@ printf("Quantidade de pontos turísticos: \n");
 scanf("%d", &pontostur);
 
 printf("Carta 01 \n Estado: %s \n Cidade: %s \n", nome, carta);
-printf("Habitantes: %f \n Área: %f km² \n", populacao, area);
+printf("Habitantes: %d \n Área: %f km² \n", populacao, area);
 printf("Número de pontos turísticos: %d \n", pontostur);
 
 // CARTA 02 -------------------------------------------------------------
@@ -48,7 +50,7 @@ printf("Nome da cidade: \n");
 scanf("%s", &carta);
 
 printf("População: \n");
-scanf("%f", &populacao);
+scanf("%d", &populacao);
 
 printf("Área em KM2: \n");
 scanf("%f", &area);
@@ -57,7 +59,7 @@ printf("Quantidade de pontos turísticos: \n");
 scanf("%d", &pontostur);
 
 printf("Carta 02 \n Estado: %s \n Cidade: %s \n", nome, carta);
-printf("Habitantes: %f \n Área: %f km² \n", populacao, area);
+printf("Habitantes: %d \n Área: %f km² \n", populacao, area);
 printf("Número de pontos turísticos: %d \n", pontostur);
